requirements.txt
torch==1.11.0
torchaudio==0.12.0
numpy
tqdm
pyaml
matplotlib
pandas

wandb

pillow
gdown